+++ mdvis/src/ui/ui_ext.h
@@ -16,11 +16,11 @@ public:
 	static void LabelMul(float x, float y, float sz, const std::string& s);
 	static std::string EditText(float x, float y, uint w, const std::string& title, const std::string& val, bool enabled = true, Vec4 col = white(1, 0.5f));
 	static std::string EditPass(float x, float y, uint w, const std::string& title, const std::string& val, bool enabled = true, Vec4 col = white(1, 0.5f));
-	static float Slider(float x, float y, uint w, const std::string& title, float a, float b, float t);
-	static float Slider(float x, float y, uint w, const std::string& title, float a, float b, float t, const std::string& lbl);
-	static float Slider(float x, float y, uint w, float a, float b, float t);
-	static void Color(float x, float y, uint w, const std::string& title, Vec4& col);
-	static void File(float x, float y, uint w, const std::string& title, const std::string& fl, filecallback func);
+	static float Slider(float x, float y, float w, const std::string& title, float a, float b, float t);
+	static float Slider(float x, float y, float w, const std::string& title, float a, float b, float t, const std::string& lbl);
+	static float Slider(float x, float y, float w, float a, float b, float t);
+	static void Color(float x, float y, float w, const std::string& title, Vec4& col);
+	static void File(float x, float y, float w, const std::string& title, const std::string& fl, filecallback func);
 	static MOUSE_STATUS Button2(float x, float y, float w, const std::string& s, const Texture& tex, Vec4 col = white(1, 0.4f), Vec4 col2 = white());
 	static void Dropdown(float x, float y, float w, const std::string& title, const Popups::DropdownItem& data);
 	static void Dropdown(float x, float y, float w, const Popups::DropdownItem& data, std::function<void()> func = nullptr, std::string label = "\1", Vec4 col = white());
